Pushing Product & ProductDisplay in TS
