+++ frontend/src/components/Product/Product.tsx
@@ -1,15 +1,33 @@
 import "./style.css";
 import React from "react";
-import Logo from "../../assets/logo.png";
+import { motion } from "framer-motion";
 
-export const Product = () => {
+export const Product = ({ product }: any) => {
+  // Check more information about framer-motion
+  // https://www.framer.com/docs/animation/
+  const animationVariants = {
+    hidden: { opacity: 0, y: -20 },
+    show: { opacity: 1, y: 0 },
+  };
+
   return (
-    <div className={"productWrapper"}>
-      <img src={Logo} alt={"coffee"} />
-      <h2>Coffee title</h2>
-      <p> 420g </p>
-      <h3>45 DKK</h3>
-      <button>Details</button>
-    </div>
+    <motion.div className={"productWrapper"} variants={animationVariants}>
+      {product.discountAmount !== "no" && <p className={"saleTag"}>ON SALE</p>}
+
+      <img src={product.productImage} alt={"coffee"} />
+      <h2>{product.productName}</h2>
+      <p> {product.productWeight}</p>
+      <h4>{product.productCategories.roast}</h4>
+      <h4>{product.productCategories.brand}</h4>
+      <h3>{product.productPrice} DKK</h3>
+
+      {product.discountAmount !== "no" && (
+        <h4 className={"discountPrice"}>
+          {product.productPrice + parseInt(product.discountAmount)} DKK
+        </h4>
+      )}
+
+      <a href={"/product/" + product.productId}>Details</a>
+    </motion.div>
   );
 };
