+++ frontend/src/components/ProductDisplay/index.ts
@@ -1,0 +1,1 @@
+export { ProductDisplay } from "./ProductDisplay";
